Cleaned up some files, a bit of refactoring. I really like using these new PresentationControls tools from DREI.
